コントローラーから切り出し
