+++ app/views/plans/index.html.erb
@@ -8,7 +8,7 @@
   <div class="side">
     <div class="introduction">
       <div class="show-area">
-        <%= link_to user_path(@user.id) do %>
+        <%= link_to user_path(current_user.id) do %>
           <%= image_tag current_user.image, class: "introduction-image" %> <br>
         <% end %>
         <div class="user-name"><%= current_user.nickname %></div>
